+++ CHANGELOG.md
@@ -7,8 +7,8 @@
   the `setUp()` method.
 * the `HelperTestCase` provides the `$Helper` property through the `__get()` magic method, and no longer through the
   `setUp()` method;
-* `assertExitWithError()` and `assertExitWithSuccess()` methods provided by `ConsoleIntegrationTestTrait` had been
-  deprecated and have been removed.
+* `getMockForComponent()` and `getMockForController()` methods provided by `MockTrait` and `assertExitWithError()` and
+  `assertExitWithSuccess()` methods provided by `ConsoleIntegrationTestTrait` had been deprecated and have been removed.
 
 ## 2.22 branch
 ### 2.22.2